fix: Add cta eligibility checks to cta-utils and apply to inline-cta-api
